refactor: moved merge conceptual out
